Refactor : user와 manage 구분 하여 디렉터리 분리
