+++ src/main/java/com/example/feedprep/common/security/config/SecurityConfig.java
@@ -48,8 +48,8 @@ public class SecurityConfig {
                             "/portone-webhook").permitAll()
                         .requestMatchers( "/notifications/subscribe").permitAll()
                     .requestMatchers(HttpMethod.OPTIONS, "/**").permitAll()
-                    .requestMatchers("/users/admin/authority").hasRole("ADMIN")
-                    .requestMatchers("/users/tutor").hasAnyRole("PENDING_TUTOR","APPROVED_TUTOR","ADMIN")
+                    .requestMatchers("/manage/admin").hasRole("ADMIN")
+                    .requestMatchers("/manage/tutor").hasAnyRole("PENDING_TUTOR","APPROVED_TUTOR","ADMIN")
 
                     .anyRequest().authenticated()
                 )
