+++ src/main/java/com/example/feedprep/domain/manage/service/AdminService.java
@@ -1,9 +1,8 @@
-package com.example.feedprep.domain.user.service;
+package com.example.feedprep.domain.manage.service;
 
 import com.example.feedprep.common.message.dto.response.TutorMessageResponseDdo;
 import com.example.feedprep.domain.techstack.dto.CreateTechStackRequestDto;
-import com.example.feedprep.domain.user.dto.response.ApproveTutorResponseDto;
-import com.example.feedprep.domain.user.dto.response.TutorResponseDto;
+import com.example.feedprep.domain.manage.dto.ApproveTutorResponseDto;
 
 public interface AdminService {
