+++ src/main/java/com/example/feedprep/domain/manage/service/AdminServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.feedprep.domain.user.service;
+package com.example.feedprep.domain.manage.service;
 
 import com.example.feedprep.common.exception.base.CustomException;
 import com.example.feedprep.common.exception.enums.ErrorCode;
@@ -8,7 +8,7 @@ import com.example.feedprep.domain.notification.service.NotificationService;
 import com.example.feedprep.domain.techstack.dto.CreateTechStackRequestDto;
 import com.example.feedprep.domain.techstack.entity.TechStack;
 import com.example.feedprep.domain.techstack.repository.TechStackRepository;
-import com.example.feedprep.domain.user.dto.response.ApproveTutorResponseDto;
+import com.example.feedprep.domain.manage.dto.ApproveTutorResponseDto;
 import com.example.feedprep.domain.user.entity.User;
 import com.example.feedprep.domain.user.enums.UserRole;
 import com.example.feedprep.domain.user.repository.UserRepository;
