+++ src/test/java/com/example/feedprep/domain/user/service/AdminServiceImplTest.java
@@ -8,11 +8,12 @@ import static org.mockito.Mockito.*;
 import com.example.feedprep.common.exception.base.CustomException;
 import com.example.feedprep.common.exception.enums.ErrorCode;
 import com.example.feedprep.common.message.service.TutorMessageService;
+import com.example.feedprep.domain.manage.service.AdminServiceImpl;
 import com.example.feedprep.domain.notification.service.NotificationService;
 import com.example.feedprep.domain.techstack.dto.CreateTechStackRequestDto;
 import com.example.feedprep.domain.techstack.entity.TechStack;
 import com.example.feedprep.domain.techstack.repository.TechStackRepository;
-import com.example.feedprep.domain.user.dto.response.ApproveTutorResponseDto;
+import com.example.feedprep.domain.manage.dto.ApproveTutorResponseDto;
 import com.example.feedprep.domain.user.entity.User;
 import com.example.feedprep.domain.user.enums.UserRole;
 import com.example.feedprep.domain.user.repository.UserRepository;
